requirements.txt
discord.py>=1.7.3
python-decouple>=3.1
setuptools>=58.0.4